013: The DiscogsSearchViewController now shows its results in itself, not in the separate DiscogsSearchResultsViewController.
