+++ SwiftDiscogsApp/Search/DiscogsSearchViewController.swift
@@ -1,44 +1,67 @@
 //  Copyright © 2018 Poikile Creations. All rights reserved.
 
 import PromiseKit
+import Stylobate
 import SwiftDiscogs
 import UIKit
 
 /// Allows the user to search the Discogs database for artists, releases, and
 /// labels.
-open class DiscogsSearchViewController: UIViewController, DiscogsProvider {
+open class DiscogsSearchViewController: OutlettedController, UISearchResultsUpdating, DiscogsProvider {
 
-    // MARK: Outlets
+    // MARK: Public Properties
+
+    /// The Discogs client.
+    open var discogs: Discogs? = /* MockDiscogs() */ DiscogsClient.singleton
+
+    /// The data model that holds the search results.
+    open var searchResultsModel: DiscogsSearchResultsModel? {
+        return model as? DiscogsSearchResultsModel
+    }
 
     /// The object responsible for managing the on-screen contents of the
     /// Discogs search, which helps keep the size of this view controller down.
+    /// The root view in the storyboard MUST have `DiscogsSearchView` as its
+    /// custom class!
     open var searchView: DiscogsSearchView? {
         return view as? DiscogsSearchView
     }
 
-    // MARK: Properties
+    // MARK: Private Properties
 
-    /// The Discogs client.
-    lazy public var discogs: Discogs? = {
-        return /* MockDiscogs() */ DiscogsClient.singleton
-    }()
+    /// The search results. Changes trigger a reloading of the table and/or
+    /// collection.
+    open var results: [DiscogsSearchResult]? {
+        didSet {
+            searchResultsModel?.results = results
+            display?.refresh()
+        }
+    }
 
     /// The `UISearchController`. It displays its results in a separate view
     /// controller (the `DiscogsSearchResultsController`).
     fileprivate lazy var searchController: UISearchController = {
-        // Load the search results controller from the storyboard.
-        let bundle = Bundle(for: type(of: self))
-        let storyboard = UIStoryboard(name: "Main", bundle:  bundle)
-        let searchResultsController
-            = storyboard.instantiateViewController(withIdentifier: "discogsSearchResults") as? DiscogsSearchResultsController
-        searchResultsController?.discogs = discogs
-        let searchController = UISearchController(searchResultsController: searchResultsController)
-        searchController.searchResultsUpdater = searchResultsController
+        let searchController = UISearchController(searchResultsController: nil)
+        searchController.searchResultsUpdater = self
 
         return searchController
     }()
     
     // MARK: UIViewController
+
+    open override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        if let segueId = segue.identifier {
+            switch segueId {
+            case "showArtist":
+                if let navController = segue.destination as? UINavigationController,
+                    let artistViewController = navController.topViewController as? DiscogsArtistViewController {
+                    artistViewController.artistSearchResult = selectedArtistResult
+                }
+            default:
+                return
+            }
+        }
+    }
 
     override open func viewDidLoad() {
         super.viewDidLoad()
@@ -97,6 +120,29 @@ open class DiscogsSearchViewController: UIViewController, DiscogsProvider {
         searchView?.signedOut()
         navigationItem.searchController = nil
         completion?()
+    }
+
+    // MARK: UISearchResultsUpdating
+
+    open func updateSearchResults(for searchController: UISearchController) {
+        let searchTerms = searchController.searchBar.text ?? ""
+        let promise: Promise<DiscogsSearchResults>? = discogs?.search(for: searchTerms, type: "Artist")
+        promise?.then { [weak self] (searchResults) -> Void in
+            self?.results = searchResults.results?.filter { $0.type == "artist" }
+            }.catch { [weak self] (error) in
+                self?.results = nil
+                self?.presentAlert(for: error)
+        }
+    }
+
+    // MARK: Everything Else
+
+    fileprivate var selectedArtistResult: DiscogsSearchResult? {
+        if let indexPath = searchView?.tableView?.indexPathForSelectedRow {
+            return searchResultsModel?.results?[indexPath.row]
+        } else {
+            return nil
+        }
     }
 
 }
